+++ p2panda-stream/src/spaces/processor.rs
@@ -66,7 +66,7 @@ where
 
         let result = if let SpacesProcessorArgs::Process { msg } = input_args {
             // Process incoming event.
-            let (groups_y, space_y, events) = match self.manager.process(msg).await {
+            let (groups_y, space_y, mut events) = match self.manager.process(msg).await {
                 Ok(result) => result,
                 Err(err) => return Err((input, SpacesError::SpacesManager(err.to_string()))),
             };
@@ -92,6 +92,36 @@ where
 
             if let Some(y) = space_y {
                 let space_id = y.space_id;
+
+                // Filter out application events if the author has concurrently lost their write
+                // access.
+                //
+                // This case occurs if a member published messages before receiving the control
+                // message which removed them. Their claim to have access write is still
+                // historically valid, however we have already learned that they have actually
+                // been removed.
+                //
+                // NOTE: Application _and_ control messages are partially-ordered based on their
+                // causal relationships, application messages are always emitted together with the
+                // member's (create/add) welcome message. For that reason application messages
+                // from members which have been removed at a causally later point (not
+                // concurrently) will not be affected by this filter and are correctly forwarded
+                // to the application layer.
+                if msg.is_application_message() {
+                    let mut is_member = false;
+
+                    // Check if the message author is still a member of the group from our local
+                    // perspective.
+                    for (member, access) in y.groups_y.members(y.group_id) {
+                        if member == msg.author && (access.is_write() || access.is_manage()) {
+                            is_member = true;
+                        }
+                    }
+
+                    if !is_member {
+                        events.retain(|event| !matches!(event, Event::Application { .. }));
+                    };
+                }
 
                 if let Err(err) = self
                     .store